+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.0.0 (not yet released)
 - Add async versions with custom cryptographic methods (**FEATURE**)
 - Update Microsoft.AspNetCore.Cryptography.KeyDerivation (**FEATURE**)
+- Add GetKeyDerivationFunctionEntryIdentifiers method for getting identifiers (**FEATURE**)
 
 ## Version 0.9.6 (released 2021-12-09)
 - Update LibAES-CTR (**FEATURE**)